// src/lib/api/generated/models/produceType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

/**
 * The category of the produce.
 */
export type ProduceType = typeof ProduceType[keyof typeof ProduceType];


export const ProduceType = {
  leafy_greens: 'leafy_greens',
  cruciferous: 'cruciferous',
  root_vegetables: 'root_vegetables',
  tubers: 'tubers',
  alliums: 'alliums',
  nightshades: 'nightshades',
  cucurbits: 'cucurbits',
  winter_squash: 'winter_squash',
  legumes: 'legumes',
  stalks_stems: 'stalks_stems',
  mushrooms: 'mushrooms',
  fresh_herbs: 'fresh_herbs',
  pome_fruits: 'pome_fruits',
  stone_fruits: 'stone_fruits',
  citrus: 'citrus',
  berries: 'berries',
  melons: 'melons',
  tropical_fruits: 'tropical_fruits',
  eggs: 'eggs',
  raw_honey: 'raw_honey',
  nuts_seeds: 'nuts_seeds',
  microgreens: 'microgreens',
  grains_pulses: 'grains_pulses',
} as const;