// src/lib/api/generated/models/produceStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

/**
 * The status of the produce listing
 */
export type ProduceStatus = typeof ProduceStatus[keyof typeof ProduceStatus];


export const ProduceStatus = {
  active: 'active',
  paused: 'paused',
  deleted: 'deleted',
} as const;