// src/lib/api/generated/models/produceStatusProperty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

/**
 * @nullable
 */
export type ProduceStatusProperty = typeof ProduceStatusProperty[keyof typeof ProduceStatusProperty] | null;


export const ProduceStatusProperty = {
  active: 'active',
  paused: 'paused',
  deleted: 'deleted',
} as const;